Main Function of correlation - regression and process accross all sites

This function applies the moving climate analysis to all sites listed in the seed.data data frame. It iterates over each site, performs the moving window analysis, and combines the results into a single data frame.

ATS_moving_climate(
  bio_data_all,
  climate.path,
  refday = 305,
  lastdays = 700,
  rollwin = 1,
  formula_model = formula("log.seed ~ TMEAN"),
  model_type = "lm"
)

Arguments

climate.path

A string specifying the path to the directory containing climate data files. These files should be named according to the site names.

refday

An integer representing the reference day for the rolling window analysis. Default is 305.

lastdays

An integer specifying the number of days to include in the rolling window analysis. Default is the maximum of a specified range.

rollwin

An integer specifying the rolling window size. Default is 1.

seed.data

A data frame containing seed count and site-level information. It should include a column plotname.lon.lat to specify site names and a column log.seed for seed counts.

Value

A data frame containing the results of the moving window analysis for all sites. Includes correlation coefficients, model coefficients, and other relevant statistics for each site.

Details

The function performs the following steps:

Identifies unique site names from the seed.data data frame.

Applies the site.moving.climate.analysis function to each site using the unique site names.

Combines the results from all sites into a single data frame.

Examples

# Example usage:
Fagus_seed <- data.frame(
  plotname.lon.lat = rep(c('Site1', 'Site2'), each = 10),
  log.seed = rnorm(20),
  year = rep(2000:2009, 2)
)
climate_beech_path <- 'path/to/climate/data'
result <- FULL.moving.climate.analysis(
  seed.data = Fagus_seed,
  climate.path = climate_beech_path,
  refday = 305,
  lastdays = 30,
  rollwin = 1
)
#> Error in FULL.moving.climate.analysis(seed.data = Fagus_seed, climate.path = climate_beech_path,     refday = 305, lastdays = 30, rollwin = 1): could not find function "FULL.moving.climate.analysis"