Optimize and Fit Generalized Additive Models (GAMs)

This function fits Generalized Additive Models (GAMs) to predict slope and R-squared values based on the provided data. It includes an option to optimize the number of knots used in the smooth term of the GAM. If optimization is enabled, it searches for the optimal number of knots that results in significant p-values for the model's smooth terms.

optimize_and_fit_gam(temporary, optim.k = TRUE, plots = F, k = 20)

Arguments

temporary

A data frame containing the data to fit the GAM models. It must include at least two columns: slope (the response variable for the slope model) and r_s (the response variable for the R-squared model), as well as day (the predictor variable).

optim.k

A logical flag indicating whether to optimize the number of knots (k) for the smooth term. If TRUE, the function will search for the optimal number of knots. Default is TRUE.

plots

A logical flag indicating whether to plot the fitted GAM models. If TRUE, plots of the fitted models and their residuals will be displayed. Default is FALSE.

k

An integer specifying the number of knots to use if optim.k is FALSE. Default is 20. If optim.k is TRUE, this parameter is ignored.

Value

A list containing:

slope_gam: The fitted GAM model for the slope.

rs_gam: The fitted GAM model for R-squared.

k: The number of knots used in the models.

Details

If optim.k is TRUE, the function iterates over a range of possible knot values (from 10 to 365). For each value, it fits a GAM model and checks the significance of the smooth term using k.check. The first value of k that results in significant p-values for all tests is chosen as the optimal number of knots. If no optimal value is found, the default value of k = -1 is used.

The function fits two GAM models: one for the slope and one for r_s, both with the chosen number of knots. If plots is TRUE, it plots the fitted models and their residuals using cowplot.

Examples

# Example data
set.seed(123)
temporary <- data.frame(
  day = 1:365,
  slope = rnorm(365),
  r_s = rnorm(365)
)

# Fit GAM models with automatic optimization of knots
result <- optimize_and_fit_gam(temporary, optim.k = TRUE, plots = TRUE)
#> Error in optimize_and_fit_gam(temporary, optim.k = TRUE, plots = TRUE): could not find function "optimize_and_fit_gam"

# Fit GAM models with a fixed number of knots
result_fixed_k <- optimize_and_fit_gam(temporary, optim.k = FALSE, k = 30, plots = TRUE)
#> Error in optimize_and_fit_gam(temporary, optim.k = FALSE, k = 30, plots = TRUE): could not find function "optimize_and_fit_gam"