Calculate Parameter Ranges

This function processes the input data for parameters alpha, beta, and sigma, and returns either the mean and standard deviation or the minimum and maximum values, based on the specified option.

parameter.range(
  raw.data.param.alpha,
  raw.data.param.beta,
  raw.data.param.sigma,
  option = "mean.sd"
)

Arguments

raw.data.param.alpha

Numeric vector or data for the parameter alpha.

raw.data.param.beta

Numeric vector or data for the parameter beta.

raw.data.param.sigma

Numeric vector or data for the parameter sigma.

option

Character. Either "mean.sd" to return the mean and standard deviation, or "min.max" to return the minimum and maximum values. Default is "mean.sd".

Value

A list containing the processed values for:

alpha: Processed values for the alpha parameter.

beta: Processed values for the beta parameter.

sigma: Processed values for the sigma parameter.

Details

This function takes in raw data for parameters alpha, beta, and sigma. Depending on the option argument, it calculates either:

the minimum and maximum values (min.max), or

the mean and standard deviation (mean.sd). The function uses helper functions get.min.max() and get.mean.sd() to perform these calculations. If an invalid option is provided, the function will stop and return an error message.

Examples

# Example with mean and standard deviation option
alpha <- rnorm(100)
beta <- rnorm(100)
sigma <- rnorm(100)
param_ranges <- parameter.range(alpha, beta, sigma, option = "mean.sd")
print(param_ranges)
#> $alpha
#> [1] 0.06589024 1.05975656
#> 
#> $beta
#> [1] 0.2401778 1.0574831
#> 
#> $sigma
#> [1] -0.01655234  1.08101049
#> 

# Example with min and max option
param_ranges_min_max <- parameter.range(alpha, beta, sigma, option = "min.max")
print(param_ranges_min_max)
#> $alpha
#> [1] -2.507918  2.691714
#> 
#> $beta
#> [1] -1.925145  2.684859
#> 
#> $sigma
#> [1] -2.549343  2.832226
#>