Format Climate Data for Moving Windows analysis

This function processes climate data to create a historical moving window of climate variables based on a reference day and a period going back in time. It adjusts for leap years, applies a rolling average to the specified climate variable, and prepares the data for analysis over past periods. Here you could adjust the rolling window to more than 1day (for example if you want average temperature over 7 days, or more or less), but in our method comparison, we set the default to 1

reformat_climate_backtothepast(
  yearsref = 2000,
  climate_data = climate_data,
  yearneed = 2,
  refday = 274,
  lastdays = 1095,
  rollwin = 1,
  covariates.of.interest = "temperature.degree",
  align.moving = "right"
)

Arguments

yearsref

Numeric. The reference year from which to start the calculations. Default is 2000.

climate_data

Data frame. The climate_data dataset containing variables such as temperature and date information.

yearneed

Numeric. The number of years needed to adjust for fruit maturation time. Default is 2. But the idea is that if you want to look more than 1095 days to the past, you need to adjust here

refday

Numeric. The reference day of the year (DOY) from which the moving window calculation begins. Default is 274.

lastdays

Numeric. The total number of days to go back from the reference day. Default is 1095 (3 years).

rollwin

Numeric. The window size for the rolling average calculation. Default is 1.

covariates.of.interest

Character. The name of the climate_data variable to apply the rolling average to. Default is 'temperature.degree' because it is my variable temperature name.

Value

A data frame containing:

LONGITUDE

Longitude of the location.

LATITUDE

Latitude of the location.

year

The year of each record.

date

The date of each record.

yday

Day of Year (DOY) for each date.

days.reversed

Number of days going back from the reference day.

rolling_avg_clim

Rolling average of the specified climate_data variable.

Details

The function performs the following tasks:

Checks if the specified variable exists in the climate_data dataset.

Adjusts the reference day if the reference year is a leap year.

Filters the climate_data data to include only the relevant years.

Creates a sequence of days going back from the reference day.

Applies a rolling average to the specified climate_data variable.

Returns a data frame with the processed climate data, including the rolling average.

If the specified covariates.of.interest is not found in the climate_data dataset, a warning is issued, and the function returns NULL.

Examples

# Example usage of the reformat.climate.backtothepast function
processed_climate_data <- reformat_climate_backtothepast(yearsref = 2000,
                                                         climate_data = climate_data,
                                                         yearneed = 2,
                                                         refday = 274,
                                                         lastdays = 1095,
                                                         rollwin = 7,
                                                         covariates.of.interest = 'temperature.degree')
#> Error: object 'climate_data' not found
head(processed_climate_data)
#> Error: object 'processed_climate_data' not found