Perform Moving Window Analysis for a Specific Site

This function performs a moving window analysis on seed and climate data for a specific site. It involves loading and formatting climate data, applying rolling window functions or not if set rolling at 1 - as done in our study, and then running a moving window analysis to investigate the relationship between seed count and climate variables.

runing_daily_relationship(
  bio_data,
  climate_data,
  lastdays,
  formula_model,
  refday = 305,
  yearneed = 2,
  model_type = "lm",
  rollwin = 1
)

Arguments

bio_data

A data frame containing seed count and site-level information, including plotname.lon.lat and log.seed.

lastdays

An integer specifying the number of days to consider in the rolling window analysis.

formula_model

A formula specifying the model to fit for each moving window. Default is formula('log.seed~rolling_avg_tmean').

site.name

A string specifying the site name for which to perform the analysis. This should match entries in the plotname.lon.lat column of bio_data.

climate.path

A string specifying the path to the directory containing climate data files. These files should include the site name in their names.

Value

A data frame containing the results of the moving window analysis for the specified site. Includes correlation coefficients, model coefficients, and other relevant statistics.

Details

The function performs the following steps:

Loads and formats the biological data for the specified site.

Loads and formats the climate data, including scaling the climate variables.

Defines the period for rolling climate data based on a specified number of years.

Applies a rolling window function to the climate data for each year in the defined period.

Runs a moving window analysis using the runing.movingwin.analysis function.

Examples

# Example usage:
site_name <- 'Site1'
Fagus_seed <- data.frame(
  plotname.lon.lat = rep(c('Site1', 'Site2'), each = 10),
  log.seed = rnorm(20),
  year = rep(2000:2009, 2)
)
climate_beech_path <- 'path/to/climate/data'
last_days <- 30
my_form <- formula('log.seed~rolling_avg_tmean')

result <- site.moving.climate.analysis(
  site.name = site_name,
  bio_data = Fagus_seed,
  climate.path = climate_beech_path,
  lastdays = last_days,
  formula_model = my_form
)
#> Error in site.moving.climate.analysis(site.name = site_name, bio_data = Fagus_seed,     climate.path = climate_beech_path, lastdays = last_days,     formula_model = my_form): could not find function "site.moving.climate.analysis"